% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{coins_all}
\alias{coins_all}
\title{Add COinS To All Files Of A Directory}
\usage{
coins_all(dir = "content")
}
\arguments{
\item{dir}{String. Providing the path to the directory. \code{coins_all}
looks recursevly at all files under the directory provided by the path
string.}
}
\value{
Invisible empty list.
}
\description{
\code{coins_all} looks for all files under the directory and calls with the
appropiate path string for each file \code{coins_this}, which is doing all the
hard work: extracting the information from frontmatter and config file,
generating the COinS string and adding it to the end of the file.
}
\details{
Two conditions prevent calling \code{coins_this()}:

\itemize{
\item File is an index file ('_index-md').
\item The YAML frontmatter contains the directive 'no_coins: yes'.
}
}
\examples{
\dontrun{
coins_all()
coins_all("content/post")}

}
