% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\docType{package}
\name{coinsR}
\alias{coinsR}
\alias{coinsR-package}
\title{The \pkg{coinsR} package}
\description{
The comprehensive documentation of this package is the book \bold{blogdown:
Creating Websites with R Markdown}
(\url{https://bookdown.org/yihui/blogdown/}).
}
\examples{
if (interactive()) coins_all()
}
